Split socks example into simple/validation
